+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Upgrade urllib3 to 2.2.3.
+- Add support for dotenv files.
 
 ## 1.1.0 (2024-05-02)